# static/swagger.json
{
    "openapi": "3.0.3",
    "info":{ 
        "title":"Tworst! - Binar Challenge Gold",
        "version":"1.0.0",
        "description": "Clean a worse tweet you have ever seen",
        "termsOfService": "https://www.github.com/haryakumuda",
        "contact": {
            "name": "Harya Kumuda Koostanto",
            "email": "haryakumuda.k@gmail.com",
            "url": "https://www.github.com/haryakumuda"
        },
        "license": {
            "name": "APACHE 2.0",
            "url": "htpps://www.apache.org/licenses/LICENSE-2.0"
        }
    },




    "servers":[
        {
            "description": "Tworst! RESTful API Server",
            "url": "/"
         }

    ],


    "externalDocs": {
        "description": "Github Code - Binar Gold Challenge",
        "url": "https://github.com/haryakumuda/Binar_Gold_Challenge"
    },



    "paths":{
        "/":{
            "get":{
                "summary": "Welcome Page",
                "description": "Welcome to Tworst!",
                "responses": {
                    "200":{
                        "description": "Success"
                    },
                    "400":{
                        "description": "Misunderstood"
                    },
                    "401":{
                        "description": "Unauthorized"
                    },
                    "404":{
                        "description": "Not Found"
                    },
                    "500":{
                        "description": "Server Error"
                    }
                }
            }
        },

        "/tweet":{
            "get":{
                "summary": "Get all of Tworst list",
                "description": "Thousand of Tworst list",
                "responses": {
                    "200":{
                        "description": "Success"
                    },
                    "400":{
                        "description": "Misunderstood"
                    },
                    "401":{
                        "description": "Unauthorized"
                    },
                    "404":{
                        "description": "Not Found"
                    },
                    "500":{
                        "description": "Server Error"
                    }
                }
            },
            "post":{
                "summary": "Post a text",
                "description": "Post a text file to clean",
                "requestBody":{
                    "required": true,
                    "content": {
                        "multipart/form-data":{
                            "schema":{
                                "type": "object",
                                "properties": {
                                    "text":{
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "description": "Success"
                    },
                    "400":{
                        "description": "Misunderstood"
                    },
                    "401":{
                        "description": "Unauthorized"
                    },
                    "404":{
                        "description": "Not Found"
                    },
                    "500":{
                        "description": "Server Error"
                    }
                }
            }
        },





        "/tweet/{tweet_id}":{
            "get":{
                "summary": "Get tweet by id",
                "description": "Get tweet by id from database",
                "parameters": [
                    {
                        "name": "tweet_id",
                        "in": "path",
                        "required": true,
                        "description": "tweet_id",
                        "schema":{
                            "type":"integer"
                        }

                    }
                ],
                "responses": {
                    "200":{
                        "description": "Success"
                    },
                    "400":{
                        "description": "Misunderstood"
                    },
                    "401":{
                        "description": "Unauthorized"
                    },
                    "404":{
                        "description": "Not Found"
                    },
                    "500":{
                        "description": "Server Error"
                    }
                }
            },
            "put":{
                "summary": "Edit existing tweet",
                "description": "Edit existing tweet in database",
                "parameters": [
                    {
                        "name": "tweet_id",
                        "in": "path",
                        "required": true,
                        "description": "tweet_id",
                        "schema":{
                            "type":"integer"
                        }

                    }
                ],
                "requestBody":{
                    "required": true,
                    "content": {
                        "multipart/form-data":{
                            "schema":{
                                "type": "object",
                                
                                "properties": {
                                    "text":{
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "description": "Success"
                    },
                    "400":{
                        "description": "Misunderstood"
                    },
                    "401":{
                        "description": "Unauthorized"
                    },
                    "404":{
                        "description": "Not Found"
                    },
                    "500":{
                        "description": "Server Error"
                    }
                }
            },
            "delete":{
                "summary": "Delete existing tweet",
                "description": "Delete existing tweet in database",
                "parameters": [
                    {
                        "name": "tweet_id",
                        "in": "path",
                        "required": true,
                        "description": "tweet_id",
                        "schema":{
                            "type":"integer"
                        }
                    }
                ],
                "requestBody":{
                    "required": true,
                    "content": {
                        "multipart/form-data":{
                            "schema":{
                                "type": "object",
                                "properties": {
                                    "text":{
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "description": "Success"
                    },
                    "400":{
                        "description": "Misunderstood"
                    },
                    "401":{
                        "description": "Unauthorized"
                    },
                    "404":{
                        "description": "Not Found"
                    },
                    "500":{
                        "description": "Server Error"
                    }
                }
            }

        },

        "/tweet/csv":{
            "post":{
                "summary": "Post a csv file",
                "description": "Post a csv file to clean",
                "requestBody":{
                    "required": true,
                    "content": {
                        "multipart/form-data":{
                            "schema":{
                                "type": "object",
                                "properties": {
                                    "file":{
                                        "type": "string",
                                        "format": "base64"
                                    }
                                }
   
                            }
                        }
                    }
                },
                "responses": {
                    "200":{
                        "description": "Success"
                    },
                    "400":{
                        "description": "Misunderstood"
                    },
                    "401":{
                        "description": "Unauthorized"
                    },
                    "404":{
                        "description": "Not Found"
                    },
                    "500":{
                        "description": "Server Error"
                    }
                }
            }
        }

    }
}